+++ contrib/examples/actions/pythonactions/isprime.py
@@ -1,12 +1,11 @@
 import math
 
+from st2actions.runners.pythonrunner import Action
 
-class PrimeChecker(object):
 
-    def __init__(self, config=None):
-        pass
-
+class PrimeCheckerAction(Action):
     def run(self, value=0):
+        self.logger.debug('value=%s' % (value))
         if math.floor(value) != value:
             raise ValueError('%s should be an integer.' % value)
         if value < 2:
@@ -17,6 +16,6 @@ class PrimeChecker(object):
         return True
 
 if __name__ == '__main__':
-    checker = PrimeChecker()
+    checker = PrimeCheckerAction()
     for i in range(0, 10):
-        print '%s : %s' % (i, checker.run(**{'value': i}))
+        print '%s : %s' % (i, checker.run(value=1))