+++ st2actions/st2actions/runners/python_action_wrapper.py
@@ -16,11 +16,12 @@
 import sys
 import json
 import argparse
-import logging as stdlib_logging
 
 from st2common import log as logging
 from st2actions import config
 from st2actions.runners.pythonrunner import Action
+from st2actions.runners.utils import get_logger_for_python_runner_action
+from st2actions.runners.utils import get_action_class_instance
 from st2common.util import loader as action_loader
 from st2common.util.config_parser import ContentPackConfigParser
 from st2common.constants.action import ACTION_OUTPUT_RESULT_DELIMITER
@@ -28,10 +29,43 @@ from st2common.service_setup import db_setup
 from st2common.services.datastore import DatastoreService
 
 __all__ = [
-    'PythonActionWrapper'
+    'PythonActionWrapper',
+    'ActionService'
 ]
 
 LOG = logging.getLogger(__name__)
+
+
+class ActionService(object):
+    """
+    Instance of this class is passed to the action instance and exposes "public"
+    methods which can be called by the action.
+    """
+
+    def __init__(self, action_wrapper):
+        logger = get_logger_for_python_runner_action(action_name=action_wrapper._class_name)
+
+        self._action_wrapper = action_wrapper
+        self._datastore_service = DatastoreService(logger=logger,
+                                                   pack_name=self._action_wrapper._pack,
+                                                   class_name=self._action_wrapper._class_name,
+                                                   api_username='action_service')
+
+    ##################################
+    # Methods for datastore management
+    ##################################
+
+    def list_values(self, local=True, prefix=None):
+        return self._datastore_service.list_values(local, prefix)
+
+    def get_value(self, name, local=True):
+        return self._datastore_service.get_value(name, local)
+
+    def set_value(self, name, value, ttl=None, local=True):
+        return self._datastore_service.set_value(name, value, ttl, local)
+
+    def delete_value(self, name, local=True):
+        return self._datastore_service.delete_value(name, local)
 
 
 class PythonActionWrapper(object):
@@ -92,38 +126,16 @@ class PythonActionWrapper(object):
         if config:
             LOG.info('Using config "%s" for action "%s"' % (config.file_path,
                                                             self._file_path))
-
-            action_instance = action_cls(config=config.config)
+            config = config.config
         else:
             LOG.info('No config found for action "%s"' % (self._file_path))
-            action_instance = action_cls(config={})
+            config = None
 
-        # Setup action_instance proeprties
-        action_instance.logger = self._set_up_logger(action_cls.__name__)
-        action_instance.datastore = DatastoreService(logger=action_instance.logger,
-                                                     pack_name=self._pack,
-                                                     class_name=action_cls.__name__,
-                                                     api_username="action_service")
-
+        action_service = ActionService(action_wrapper=self)
+        action_instance = get_action_class_instance(action_cls=action_cls,
+                                                    config=config,
+                                                    action_service=action_service)
         return action_instance
-
-    def _set_up_logger(self, action_name):
-        """
-        Set up a logger which logs all the messages with level DEBUG
-        and above to stderr.
-        """
-        logger_name = 'actions.python.%s' % (action_name)
-        logger = logging.getLogger(logger_name)
-
-        console = stdlib_logging.StreamHandler()
-        console.setLevel(stdlib_logging.DEBUG)
-
-        formatter = stdlib_logging.Formatter('%(name)-12s: %(levelname)-8s %(message)s')
-        console.setFormatter(formatter)
-        logger.addHandler(console)
-        logger.setLevel(stdlib_logging.DEBUG)
-
-        return logger
 
 
 if __name__ == '__main__':